break up nav links added next and prev helpers
