+++ alethio/blocks.go
@@ -7,14 +7,16 @@ import (
 
 type BlocksService service
 
+type NavLinks struct {
+	Next string `json:"next"`
+	Prev string `json:"prev"`
+}
+
 // GetAllBlocks - structure when asking for all blocks
 type GetAllBlocks struct {
-	Data  []Block `json:"data"`
-	Links struct {
-		Next string `json:"next"`
-		Prev string `json:"prev"`
-	} `json:"links"`
-	Meta struct {
+	Data  []Block  `json:"data"`
+	Links NavLinks `json:"links"`
+	Meta  struct {
 		Count       int `json:"count"`
 		LatestBlock struct {
 			Number            int    `json:"number"`
@@ -164,85 +166,8 @@ type LogEntries struct {
 			Self string `json:"self"`
 		} `json:"links"`
 	} `json:"data"`
-	Links struct {
-		Next string `json:"next"`
-		Prev string `json:"prev"`
-	} `json:"links"`
-	Meta struct {
-		Count       int `json:"count"`
-		LatestBlock struct {
-			Number            int    `json:"number"`
-			BlockCreationTime int    `json:"blockCreationTime"`
-			BlockHash         string `json:"blockHash"`
-		} `json:"latestBlock"`
-		Page struct {
-			HasNext bool `json:"hasNext"`
-			HasPrev bool `json:"hasPrev"`
-		} `json:"page"`
-	} `json:"meta"`
-}
-
-// EtherTransfers - ether transfer array structure for the /block/{blockhash}/etherTransfers API endpoint
-type EtherTransfers struct {
-	Data []struct {
-		Type       string `json:"type"`
-		ID         string `json:"id"`
-		Attributes struct {
-			BlockCreationTime int    `json:"blockCreationTime"`
-			Cursor            string `json:"cursor"`
-			Fee               string `json:"fee"`
-			GlobalRank        []int  `json:"globalRank"`
-			Total             string `json:"total"`
-			TransferType      string `json:"transferType"`
-			Value             string `json:"value"`
-		} `json:"attributes"`
-		Relationships struct {
-			Block struct {
-				Data struct {
-					Type string `json:"type"`
-					ID   string `json:"id"`
-				} `json:"data"`
-				Links struct {
-					Related string `json:"related"`
-				} `json:"links"`
-			} `json:"block"`
-			ContractMessage struct {
-				Data interface{} `json:"data"`
-			} `json:"contractMessage"`
-			FeeRecipient struct {
-				Data struct {
-					Type string `json:"type"`
-					ID   string `json:"id"`
-				} `json:"data"`
-				Links struct {
-					Related string `json:"related"`
-				} `json:"links"`
-			} `json:"feeRecipient"`
-			From struct {
-				Data interface{} `json:"data"`
-			} `json:"from"`
-			To struct {
-				Data struct {
-					Type string `json:"type"`
-					ID   string `json:"id"`
-				} `json:"data"`
-				Links struct {
-					Related string `json:"related"`
-				} `json:"links"`
-			} `json:"to"`
-			Transaction struct {
-				Data interface{} `json:"data"`
-			} `json:"transaction"`
-		} `json:"relationships"`
-		Links struct {
-			Self string `json:"self"`
-		} `json:"links"`
-	} `json:"data"`
-	Links struct {
-		Next string `json:"next"`
-		Prev string `json:"prev"`
-	} `json:"links"`
-	Meta struct {
+	Links NavLinks `json:"links"`
+	Meta  struct {
 		Count       int `json:"count"`
 		LatestBlock struct {
 			Number            int    `json:"number"`
@@ -332,11 +257,8 @@ type TokenTransfers struct {
 			Self string `json:"self"`
 		} `json:"links"`
 	} `json:"data"`
-	Links struct {
-		Next string `json:"next"`
-		Prev string `json:"prev"`
-	} `json:"links"`
-	Meta struct {
+	Links NavLinks `json:"links"`
+	Meta  struct {
 		Count       int `json:"count"`
 		LatestBlock struct {
 			Number            int    `json:"number"`
