+++ alethio/contract_messages.go
